+++ WordPress/src/main/java/org/wordpress/android/ui/compose/components/menu/dropdown/JetpackDropdownMenu.kt
@@ -1,21 +1,40 @@
 package org.wordpress.android.ui.compose.components.menu.dropdown
 
 import android.content.res.Configuration
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material.ContentAlpha
 import androidx.compose.material.Divider
+import androidx.compose.material.Icon
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.LocalTextStyle
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.MenuDefaults
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import me.saket.cascade.CascadeColumnScope
 import me.saket.cascade.CascadeDropdownMenu
 import org.wordpress.android.R
 import org.wordpress.android.ui.compose.theme.AppTheme
@@ -56,6 +75,124 @@ fun JetpackDropdownMenu(
             }
         }
     }
+}
+
+@Composable
+private fun Item(
+    element: MenuElementData.Item,
+    onMenuItemClick: (MenuElementData) -> Unit,
+) {
+    val enabledContentColor = MenuColors.itemContentColor()
+    val disabledContentColor = enabledContentColor.copy(alpha = ContentAlpha.disabled)
+    DropdownMenuItem(
+        modifier = Modifier
+            .background(MenuColors.itemBackgroundColor()),
+        onClick = {
+            onMenuItemClick(element)
+            element.onClick()
+        },
+        colors = MenuDefaults.itemColors(
+            textColor = enabledContentColor,
+            leadingIconColor = enabledContentColor,
+            trailingIconColor = enabledContentColor,
+            disabledTextColor = disabledContentColor,
+            disabledLeadingIconColor = disabledContentColor,
+            disabledTrailingIconColor = disabledContentColor,
+        ),
+        text = {
+            Text(
+                text = element.text,
+                style = MaterialTheme.typography.bodyLarge,
+                overflow = TextOverflow.Ellipsis,
+                fontWeight = FontWeight.Normal,
+                maxLines = 1,
+            )
+        },
+        leadingIcon = if (element.leadingIcon != NO_ICON) {
+            {
+                Icon(
+                    painter = painterResource(id = element.leadingIcon),
+                    contentDescription = null,
+                )
+            }
+        } else null,
+    )
+}
+
+@Composable
+private fun CascadeColumnScope.SubMenuHeader(
+    modifier: Modifier = Modifier,
+    contentPadding: PaddingValues = PaddingValues(10.5.dp),
+    text: @Composable (() -> Unit)? = null,
+) {
+    Row(
+        modifier = modifier
+            .background(MenuColors.itemBackgroundColor())
+            .fillMaxWidth()
+            .clickable(enabled = hasParentMenu, role = Role.Button) {
+                if (!isNavigationRunning) {
+                    cascadeState.navigateBack()
+                }
+            }
+            .padding(contentPadding),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        CompositionLocalProvider(
+            LocalTextStyle provides MaterialTheme.typography.bodyLarge
+        ) {
+            if (this@SubMenuHeader.hasParentMenu) {
+                Image(
+                    painter = painterResource(R.drawable.ic_arrow_left_white_24dp),
+                    contentDescription = null,
+                    colorFilter = ColorFilter.tint(MenuColors.itemContentColor()),
+                )
+            }
+            Box(Modifier.weight(1f)) {
+                text?.invoke()
+            }
+        }
+    }
+}
+
+@Composable
+private fun CascadeColumnScope.SubMenu(
+    element: MenuElementData.SubMenu,
+    onMenuItemClick: (MenuElementData) -> Unit,
+) {
+    val enabledContentColor = MenuColors.itemContentColor()
+    val disabledContentColor = enabledContentColor.copy(alpha = ContentAlpha.disabled)
+    DropdownMenuItem(
+        modifier = Modifier
+            .background(MenuColors.itemBackgroundColor()),
+        colors = MenuDefaults.itemColors(
+            textColor = enabledContentColor,
+            leadingIconColor = enabledContentColor,
+            trailingIconColor = enabledContentColor,
+            disabledTextColor = disabledContentColor,
+            disabledLeadingIconColor = disabledContentColor,
+            disabledTrailingIconColor = disabledContentColor,
+        ),
+        text = {
+            Text(
+                text = element.text,
+                style = MaterialTheme.typography.bodyLarge,
+                overflow = TextOverflow.Ellipsis,
+                fontWeight = FontWeight.Normal,
+                maxLines = 1,
+            )
+        },
+        children = {
+            element.children.forEach {
+                when (it) {
+                    is MenuElementData.SubMenu -> SubMenu(it, onMenuItemClick)
+                    is MenuElementData.Item -> Item(it, onMenuItemClick)
+                }
+            }
+        },
+        childrenHeader = {
+            SubMenuHeader()
+        }
+    )
 }
 
 @Preview
